Filling the release version field on conversion
When converting from a Metadata to a V3Package we need to write the
current time in UTC as the number of seconds from the Unix Epoch. This
means that we need to convert ReleaseVersion to use a Long instead of a
Int. This is okay as JSON numberse are magical.

+++ cosmos-json/src/main/scala/com/mesosphere/universe/v3/circe/Decoders.scala
@@ -49,7 +49,7 @@ object Decoders {
 
   implicit val decodePackageDefinitionReleaseVersion: Decoder[PackageDefinition.ReleaseVersion] =
     Decoder.instance[PackageDefinition.ReleaseVersion] { (c: HCursor) =>
-      c.as[Int].map(PackageDefinition.ReleaseVersion(_)).flatMap {
+      c.as[Long].map(PackageDefinition.ReleaseVersion(_)).flatMap {
         case Return(v) => Xor.Right(v)
         case Throw(e) => Xor.Left(DecodingFailure(e.getMessage, c.history))
       }
